### 2022-02-02 #502 @assignUser
octolog
**Reviewers:**

Submitting Author Name: Jacob Wujciak-Jens
Submitting Author Github Handle: @assignUser
Repository:  https://github.com/assignUser/octolog/
Version submitted:
Submission type: Standard
Version accepted: TBD
---

```
Package: octolog
Title: Better 'Github Action' Logging
Version: 0.1.1.9000
    person("Jacob", "Wujciak-Jens", , "jacob@wujciak.de", role = c("aut", "cre"),
Description: Provides an API for 'Github' <https://github.com> workflow
    commands and makes it possible to signal conditions from within R that
    will create annotations when used within 'Github Actions'
    <https://github.com/features/actions> but raise normal R conditions
    when used interactively.
    https://assignuser.github.io/octolog/
    cli (>= 3.0.0),
    fs,
    glue,
    openssl,
    rlang (>= 1.0.0),
    withr
    covr,
    mockery (>= 0.4.2.9000),
    r-lib/mockery
Config/testthat/parallel: true

```

Octolog aims to make it easier for R author's to use the full power of Github Actions as a CI service.

Using CI is recognized as an important part in the development process of (scientific) R software and octlog aims to make this easier and more powerful.

As far as I know there are none.

https://github.com/ropensci/software-review/issues/501
