### 2022-02-08 b566d41
make openssl optional backwardscompatibility with rlang
