+++ tests/testthat/_snaps/commands.md
@@ -100,9 +100,8 @@
       octo_start_group(c("error", "too much"))
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::Group `name` must be length 1!
-    Condition
-      Error in `octo_start_group()`:
-      ! Group `name` must be length 1!
+    Error <rlang_error>
+      Group `name` must be length 1!
 
 ---
 
@@ -114,9 +113,8 @@
       octo_start_group(c("error", "too much"))
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::Group `name` must be length 1!
-    Condition
-      [1m[33mError[39m in [1m[1m[30m[47m`octo_start_group()`[49m[39m:[22m
-      [33m![39m [1m[22mGroup [30m[47m`name`[49m[39m must be length 1!
+    Error <rlang_error>
+      [1m[22mGroup [30m[47m`name`[49m[39m must be length 1!
 
 ---
 
@@ -128,9 +126,8 @@
       octo_start_group(c("error", "too much"))
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::Group `name` must be length 1!
-    Condition
-      Error in `octo_start_group()`:
-      ! Group `name` must be length 1!
+    Error <rlang_error>
+      Group `name` must be length 1!
 
 ---
 
@@ -142,9 +139,8 @@
       octo_start_group(c("error", "too much"))
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::Group `name` must be length 1!
-    Condition
-      [1m[33mError[39m in [1m[1m[30m[47m`octo_start_group()`[49m[39m:[22m
-      [33m![39m [1m[22mGroup [30m[47m`name`[49m[39m must be length 1!
+    Error <rlang_error>
+      [1m[22mGroup [30m[47m`name`[49m[39m must be length 1!
 
 ---
 
@@ -252,9 +248,8 @@
       octo_mask_value(c(1, 2))
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::You can only mask one value at a time.
-    Condition
-      Error in `octo_mask_value()`:
-      ! You can only mask one value at a time.
+    Error <rlang_error>
+      You can only mask one value at a time.
 
 ---
 
@@ -262,9 +257,8 @@
       octo_mask_envvar(c("VAR1", "VAR2"))
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::You can only mask one envvar at a time.
-    Condition
-      Error in `octo_mask_envvar()`:
-      ! You can only mask one envvar at a time.
+    Error <rlang_error>
+      You can only mask one envvar at a time.
 
 ---
 
@@ -272,9 +266,8 @@
       octo_mask_envvar("SECRET_VAR")
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::The envvar `SECRET_VAR` does not exists!
-    Condition
-      Error in `octo_mask_envvar()`:
-      ! The envvar `SECRET_VAR` does not exists!
+    Error <rlang_error>
+      The envvar `SECRET_VAR` does not exists!
 
 ---
 
@@ -294,9 +287,8 @@
       octo_mask_value(c(1, 2))
     Output
       ::error file=universe.R,line=23,endLine=42,col=3,endCol=27::You can only mask one value at a time.
-    Condition
-      [1m[33mError[39m in [1m[1m[30m[47m`octo_mask_value()`[49m[39m:[22m
-      [33m![39m [1m[22mYou can only mask one value at a time.
+    Error <rlang_error>
[…]
