### 2022-02-08 3a4cb1c
make openssl optional backwardscompatibility with rlang
