### 2022-02-08 9badebf
exclude snaps from eof hook
