+++ README.Rmd
@@ -17,7 +17,7 @@ output:
 
 Octolog provides a complete[^1] API for Github [workflow commands](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions) that makes it easy to create powerful Github Actions that create readable logs. 
 
- and makes it possible to signal conditions from within R that will create annotations when used within Github Actions but raise normal R conditions
+Additionally it is possible to signal conditions with octolog from R that will create annotations when used within Github Actions but raise normal R conditions
 when used interactively.
 ![Annotations on files of a PR](man/figures/error.png)
 
