+++ dev/pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   octolog: octolog.html
-last_built: 2022-02-02T16:52Z
+last_built: 2022-02-08T15:51Z
 urls:
   reference: https://assignuser.github.io/octolog/reference
   article: https://assignuser.github.io/octolog/articles
