### 2022-02-08 #502 @assignUser
Thank you @jooolia for your work so far!

I just wanted to note that my test coverage as displayed by {pkgcheck} has seemingly dropped below 75% but this is due to an issue with {covr} and the way it handles `expect_snapshot` tests: https://github.com/r-lib/covr/issues/482
The actual coverage is >90%: https://app.codecov.io/gh/assignUser/octolog
