+++ R/commands.R
@@ -112,9 +112,12 @@ octo_set_output <- function(value, name) {
 #' if untrusted output (e.g. issue titles, bodies or commit messages) needs to
 #' be logged this can be used to stop this output from running possibly
 #' malicious workflow commands. Requires the suggested package
-#' [openssl::openssl] to be installed.
+#' [openssl::openssl] to be installed, when used within Github Actions.
 #' @param token A unique token used to restart workflow command parsing.
-#' @return The `token` needed to reactivate the workflow command parsing.
+#' @return The `token` needed to reactivate the workflow command parsing. When
+#'   not on Github Actions returns a `token` not generated by `openssl` to
+#'   avoid the dependency.
+#'
 #' @examples
 #' Sys.setenv(GITHUB_ACTIONS = "true")
 #' tk <- octo_stop_commands()
@@ -124,17 +127,20 @@ octo_set_output <- function(value, name) {
 #' and the [Github Blog](https://github.blog/changelog/2020-10-01-github-actions-deprecating-set-env-and-add-path-commands/)
 #' @export
 octo_stop_commands <- function() {
-  if (!rlang::is_installed("openssl")) {
-    octo_abort(c(
+  if (on_github() && !rlang::is_installed("openssl")) {
+    octo_abort(
       paste0(
         "The package {.pkg openssl} must be installed to use",
         " this function safely!"
-      ),
-      i = "{.code install.packages('openssl')}"
-    ))
+      )
+    )
   }
 
-  token <- openssl::rand_bytes(12) %>% paste0(collapse = "")
+  if (on_github()) {
+    token <- openssl::rand_bytes(12) %>% paste0(collapse = "")
+  } else {
+    token <- basename(tempfile(""))
+  }
   glue("::stop-commands::{token}") %>% octocat()
 
   token
