+++ NEWS.md
@@ -1,7 +1,8 @@
 # octolog (development version)
 
 * octolog is now compatible with rlang >= 0.4.10
-* openssl is now optional as it is only used in `octo_stop_commands()`
+* openssl is now optional as it is only used in `octo_stop_commands()`. 
+  * An unsafe fallback is used to generate the token outside of Github Actions.
   
 * # octolog 0.1.1
 
