+++ man/octo_stop_commands.Rd
@@ -13,7 +13,9 @@ octo_start_commands(token)
 \item{token}{A unique token used to restart workflow command parsing.}
 }
 \value{
-The \code{token} needed to reactivate the workflow command parsing.
+The \code{token} needed to reactivate the workflow command parsing. When
+not on Github Actions returns a \code{token} not generated by \code{openssl} to
+avoid the dependency.
 }
 \description{
 This will stop github from processing any workflow commands until
@@ -21,7 +23,7 @@ This will stop github from processing any workflow commands until
 if untrusted output (e.g. issue titles, bodies or commit messages) needs to
 be logged this can be used to stop this output from running possibly
 malicious workflow commands. Requires the suggested package
-\link[openssl:openssl]{openssl::openssl} to be installed.
+\link[openssl:openssl]{openssl::openssl} to be installed, when used within Github Actions.
 }
 \examples{
 Sys.setenv(GITHUB_ACTIONS = "true")
