### 2022-02-18 0f067a7
fix: update tests for CLI 3.2.0
