### 2022-02-19 8e71de9
fix: pkgdown examples
