+++ NEWS.md
@@ -4,7 +4,7 @@
 * openssl is now optional as it is only used in `octo_stop_commands()`. 
   * An unsafe fallback is used to generate the token outside of Github Actions.
   
-* # octolog 0.1.1
+# octolog 0.1.1
 
 * add CITATION
 
