+++ .github/workflows/pkgdown.yaml
@@ -29,8 +29,9 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::pkgdown
-
+          extra-packages: |
+            any::pkgdown
+            local::.
       - name: Deploy package
         run: |
           git config --local user.name "$GITHUB_ACTOR"
