+++ .github/workflows/pkgcheck.yaml
@@ -17,6 +17,6 @@ jobs:
   check:
     runs-on: ubuntu-latest
     steps:
-      - uses: ropensci-review-tools/pkgcheck-action@main
+      - uses: ropensci-review-tools/pkgcheck-action@fix-pak
         with:
           summary-only: false
