+++ .github/workflows/pkgcheck.yaml
@@ -17,6 +17,6 @@ jobs:
   check:
     runs-on: ubuntu-latest
     steps:
-      - uses: ropensci-review-tools/pkgcheck-action@fix-pak
+      - uses: ropensci-review-tools/pkgcheck-action@main
         with:
           summary-only: false
