### 2022-03-14 #502 @pat-s
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

Estimated hours spent reviewing: 2,5

Overall I think the package is a useful contribution and can be of good use use in CI/CD workflows. It simplifies the combined usage of local and CI/CD execution. I will also try it out in combination with {tic} to better highlight and format log outputs! In fact, I've played around with it a bit in the `octolog` branch in {tic} (see [CI runs](https://github.com/ropensci/tic/actions)) and the condition signaling worked as expected!

Please finde some comments below of points which could possibly be improved/changed to make the package even better :)

I think the motivation and intro could profit from one or two sentences more highlighting the scope of the package and how {octolog} might improve the experience of a package author/analyst who works with GHA. I think the word "additionally" suggests that the condition signaling in R is just a "nice side effect" whereas it is actually one of the main points. To be clear: I think this sentence just needs a slight rewrite and the word "additionally" should probably be avoided to not "value down" this part.

I was wondering where the name {octolog} originates from? I couldn't find a statement about it - maybe this could be added? (I assume it originates from the "octo" being GH logo animal - but maybe I am wrong? :)

## Dependencies

I know this usually is a hot debate with many different views but let me share mine :) I see that there is a strong focus on tidyverse packages (which is not a problem per se and tidyverse packages are of course well programmed packages), yet they usually bring in additional dependencies and one has to ask themselves if it's worth it - especially on the R package side (scripting/data analysis is a different use case).
Please don't see this is a "blocker" or similar, just as a personal opinion or "food for thought" :)

I think the package could be substantially reduced WRT to package dependencies:

-   Often the magrittr pipe is only used in scripting/analysis and not within packages as it can be easily replaced

-   {rlang} is similar to {magrittr} state above, though I see that it might do some things "better" than base R checks etc. I see that mainly `rlang::trace_back()` is used and one call to `rlang::exec()`. I haven't yet made use of the former but wanted to ask which additional benefits the use of `rlang::trace_back()` brings compared to `traceback()` WRT to the package functionality?

-   I like to use {fs} in scripting but try to avoid in packages due to it's dependency on libuv. This is especially costly and (possibly annoying for unexperienced users) in CI as the system dependency first needs to be installed. This costs time during every run and I think the {fs} dependency could be easily avoided.

Avoiding all of the above would leave the package with {cli}, {glue} and {withr} as direct dependencies (in IMPORTS) which would be an improvement to me.

-   The initial example and screenshot is somewhat hard to understand for first time users. To make it fully clear I'd suggest to also add an example how the error msg would look like in an interactive example - maybe even side by side in a a table?

-   Installation: Given that {pak} is still quite unstable and non-standard WRT to package installation (and is a third-party package), I'd not use it as the "default" installation method. I'd suggest to use `install.packages()` and mention {pak} as an alternative. In addition, `remotes::` should be used instead of `devtools::`

-   The following URL in the README leads to a 404: [website](https://github.com/assignUser/octolog/blob/main/assignUser.github.io/octolog)

-   Footnote 2 states that GHA is "only free for public repos". It is also free for private repos but with a limited amount of usage (in mins)

## Get Started

-   The vignette indicates that {octolog} only works with {cli} message output and not with normal `message()` or `warning()` calls. If so, this should be highlighted more prominently at the start of the vignette or in the README

-   The example on `octo_abort()` should maybe get it's own section. Especially because "This can be useful if you want a workflow to fail (e.g. to block a PR) by throwing an error but do not want to stop the R process." is one of the selling features of {octolog}

-   The "masking" section could profit from an additional R code block showcasing the R code directly as the screenshot is somewhat hard to read. Also it might be of interested to note why these two functions differ and which one should be used in which scenario.

-   The "stopping command parsing" section could profit from a toy example. I find it quite hard to image a scenario in which the describe command could be of good use. The last two sentences "This is not intended to protect against other R code. Due to R's metaprogramming capabilities it is not possible to protect the token within the same R session." are not fully clear to me in this context. I think this section could profit from more context and a clear example :)

## Functions

-   <https://jacob.wujciak.de/octolog/reference/enable_github_colors.html> : Examples throw an error

-   Maybe group the pkgdown reference output in meaningful categories?

-   `octo_add_path()` : the description is unclear to me

-   `octo_mask_value()`: examples error

-   `octo_start_group()`: The example errors. I think this could be worth getting it's own section in a vignette?

-   `octocat()`: what is it's relation to the other cli-based functions?

## Stylistic

-   Github -\> GitHub

-   It would be great to somehow differentiate the package from normal text, either by formatting is as code or wrapping it in brackets such as {octolog}

-   Typos:

    -   README: powerful,free (whitespace)

    -   Get started vignette: using `octo_stop_commands`.Internally the function

-   Screenshots: Could possibly be of higher quality
