### 2022-06-08 #502 @annakrystalli
---

Fistly many apologies for the delays in getting the package through the review process. As for the package, given how popular GitHub Actions are becoming (and it's great to hear it works nicely for `tic` too), it's a nice and clean addition to packages in the R continuous integration space.

Thank you as well again @pat-s for your excellent review. I pretty much agree with all of @pat-s's comments, especially the one on streamlining dependencies. I think the suggestion is a smart one and second it.

So as not to repeat the same points, I focus on a few additional comments I have, primarily building on my initial points in the editor's checks as well as some mentioned by @pat-s regarding documentation.

#### Documentation:

##### Minor points

- In the pgkdown docs, there are many errors in examples of functions with the error `could not find function <function_name>`)  e.g.  `enable_github_colors`, `encode_string`, `octo_warn` and more (some of which run locally).

- When running examples locally, often `Sys.setenv(GITHUB_ACTIONS = "true")`
needed to be run. It would be good if example code cleaned up after itself with `withr::local_envvar(GITHUB_ACTIONS = "true")` like you do in tests.

##### General package documentation

While I appreciate that you point to a live example of GA runs using octolog from the README / Getting started, I feel more explanation of what is going on would be appreciated by new users.

For example, for starters, I would explicitly link to the [`test-octolog.yaml`](https://github.com/assignUser/octolog/blob/main/.github/workflows/test-octolog.yaml#L34-L89) workflow that is being run.

Then I would point the user to the lines in the workflow that are running the R code that are generating the messages by focusing on this section and explaining what is going on:
https://github.com/assignUser/octolog/blob/8e71de9077f0f6a3e5fbefb78be5c6f32c0cbf47/.github/workflows/test-octolog.yaml#L34-L89

More specifically I might explain that it's sourcing [`/.github/workflows/R/conditions.R`](https://github.com/assignUser/octolog/blob/main/.github/workflows/R/conditions.R) (which is so short I might even include it in the docs) and then running R code from the workflow yaml itself.

For me the key is to make clear not just the code that generates the output in GA, but also where they appear in R code in the repo and how they then get run through GitHub Actions

Two additional suggestions to make the usefulness of the package to the R user more concrete and easy to experiment with.

1. While your examples in `test-octolog.yaml` are comprehensive with respect to the generic types of outputs the octolog API and your package offer, I feel they could be improved in terms of helping R users visualise why it would be useful to include them. It would be good to provide some more real life examples (especially more realistic R continuous integration scenarios) and explain some of the ways in which R code might be run in GA (e.g. raw R code in the workflow yaml, sourcing R code, code being run as part of package tests, package loading).

  Perhaps you could set up a tiny R package and placing octolog code in actual function files in R. Messages could be deliberate error messages, package loading messages, package dependency checks, progress messages, basically whatever you think you might inspire users to visualise their use.

  I don't want to be prescriptive here, I'm just trying to motivate some ideas. The important point of all of this is to give a better and broader view of how they might be useful in GitHub Actions to an R user.

2. If you set up such a demo repository, you could even set it up as a template that users could copy (or if not just fork) and direct them to use it as a playground for testing out package functionality. This could get them started much quicker with experimenting with package functionality and for me greatly improve understanding of how the package works and how they might use it.

#### Minor points

##### good practice notes

There are still a couple of suggestions being picked up by `goodpractice::gp()`. I feel the ones about line length are not a huge deal but they still remain best practice for readability.

The note about using `setwd()`, again, I'm not sure it's a big deal because it's being used in tests and it seems and you clean up nicely with `withr::defer(setwd(old_wd))`. I'm guessing this is the best way to achieve what you're trying to do. If there's no other way, I think this will be fine.

#### spell check

Probably the only potentially valid typo is in the roxygen docs for `enable_github_colors`:
```
sideeffects         enable_github_colors.Rd:29
```

#### Conclusion

I know it might feel like my suggestions are overboard but I feel it's really important to consider the perspective of complete beginners, walk them through very explicitly how things work (instead of linking to examples with quite a few working parts that they need to reverse engineer) and help them visualise how the package can help them.

Very open to hearing your thoughts on this!
