+++ tests/testthat/_snaps/util.md
@@ -28,85 +28,43 @@
 
     Code
       enable_github_colors(.local_envir = env)
-    Message <cliMessage>
+    Message
       v Enabled colors!
 
 ---
 
     Code
       enable_github_colors(.local_envir = env)
-    Message <cliMessage>
+    Message
       i `R_CLI_NUM_COLORS` already set.
 
 ---
 
     Code
       disable_github_colors()
-    Message <cliMessage>
+    Message
       x Disabeled colors!
 
 # enable_github_colors [ansi]
 
     Code
       enable_github_colors(.local_envir = env)
-    Message <cliMessage>
+    Message
       [32mv[39m Enabled colors!
 
 ---
 
     Code
       enable_github_colors(.local_envir = env)
-    Message <cliMessage>
+    Message
       [36mi[39m `R_CLI_NUM_COLORS` already set.
 
 ---
 
     Code
       disable_github_colors()
-    Message <cliMessage>
+    Message
       x Disabeled colors!
-
-# enable_github_colors [unicode]
-
-    Code
-      enable_github_colors(.local_envir = env)
-    Message <cliMessage>
-      ✔ Enabled colors!
-
----
-
-    Code
-      enable_github_colors(.local_envir = env)
-    Message <cliMessage>
-      ℹ `R_CLI_NUM_COLORS` already set.
-
----
-
-    Code
-      disable_github_colors()
-    Message <cliMessage>
-      ✖ Disabeled colors!
-
-# enable_github_colors [fancy]
-
-    Code
-      enable_github_colors(.local_envir = env)
-    Message <cliMessage>
-      [32m✔[39m Enabled colors!
-
----
-
-    Code
-      enable_github_colors(.local_envir = env)
-    Message <cliMessage>
-      [36mℹ[39m `R_CLI_NUM_COLORS` already set.
-
----
-
-    Code
-      disable_github_colors()
-    Message <cliMessage>
-      ✖ Disabeled colors!
 
 # get_location_string
 
