+++ tests/testthat/_snaps/util.md
@@ -66,10 +66,52 @@
     Message
       x Disabeled colors!
 
+# enable_github_colors [unicode]
+
+    Code
+      enable_github_colors(.local_envir = env)
+    Message
+      ✔ Enabled colors!
+
+---
+
+    Code
+      enable_github_colors(.local_envir = env)
+    Message
+      ℹ `R_CLI_NUM_COLORS` already set.
+
+---
+
+    Code
+      disable_github_colors()
+    Message
+      ✖ Disabeled colors!
+
+# enable_github_colors [fancy]
+
+    Code
+      enable_github_colors(.local_envir = env)
+    Message
+      [32m✔[39m Enabled colors!
+
+---
+
+    Code
+      enable_github_colors(.local_envir = env)
+    Message
+      [36mℹ[39m `R_CLI_NUM_COLORS` already set.
+
+---
+
+    Code
+      disable_github_colors()
+    Message
+      ✖ Disabeled colors!
+
 # get_location_string
 
     Code
       get_location_string(trc)
     Output
-      file=test-util.R,line=61,endLine=61,col=3,endCol=19
+      file=test-util.R,line=64,endLine=64,col=3,endCol=19
 
