+++ R/commands.R
@@ -1,6 +1,6 @@
 #' Grouping log lines
 #'
-#' These functions make it possible to group lines in the Github Actions log.
+#' These functions make it possible to group lines in the GitHub Actions log.
 #' Groups can not be nested at this point, see this [issue](https://github.com/actions/runner/issues/802).
 #' @param name Name of the group, single line.
 #' @examples
@@ -9,7 +9,7 @@
 #' print("Log other output")
 #' octo_end_group()
 #' @export
-#' @seealso [Github Docs](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#grouping-log-lines)
+#' @seealso [GitHub Docs](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#grouping-log-lines)
 octo_start_group <- function(name) {
   if (length(name) != 1) {
     octo_abort("Group {.arg name} must be length 1!")
@@ -27,10 +27,10 @@ octo_end_group <- function() {
 #' Masking a value or envvar in the GHA log.
 #'
 #' This will mask either a `value` or an envvar and prevent them (or their
-#'  content) from showing up in the Github Actions log.
+#'  content) from showing up in the GitHub Actions log.
 #' \cr ***ATTENTION***: Currently the masking of envvar values will only take
 #' effect in the ***NEXT*** step of the workflow. Values that are masked
-#' directly are masked immediately. This is not very clear in the Github Docs
+#' directly are masked immediately. This is not very clear in the GitHub Docs
 #' but very important.
 #' @details The masking is not restricted to R output, rather it will work for
 #' any logged output. For a practical demonstration please see the
@@ -44,7 +44,7 @@ octo_end_group <- function() {
 #' [here](https://github.com/actions/runner/issues/475#issuecomment-742271143).
 #' @param value A single value to mask, coercible to string.
 #' @param name Name of the envvar to mask.
-#' @seealso [Github Docs](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#masking-a-value-in-log)
+#' @seealso [GitHub Docs](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#masking-a-value-in-log)
 #' @examples
 #' octo_mask_value("secret_token123")
 #' # The mask takes effect in the NEXT step
@@ -87,7 +87,7 @@ octo_mask_envvar <- function(name) {
 #' @param value A single line string (or coercible to character). Use
 #'   [encode_string()] to encode a numeric
 #'   or multiline string.
-#' @seealso [Github Docs](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-output-parameter)
+#' @seealso [GitHub Docs](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-output-parameter)
 #' @examples
 #' Sys.setenv(GITHUB_ACTION = "true")
 #' string <- "A multi-line \n String."
@@ -112,19 +112,19 @@ octo_set_output <- function(value, name) {
 #' if untrusted output (e.g. issue titles, bodies or commit messages) needs to
 #' be logged this can be used to stop this output from running possibly
 #' malicious workflow commands. Requires the suggested package
-#' [openssl::openssl] to be installed, when used within Github Actions.
+#' [openssl::openssl] to be installed, when used within GitHub Actions.
 #' @param token A unique token used to restart workflow command parsing.
 #' @return The `token` needed to reactivate the workflow command parsing. When
[…]
