### 2022-06-25 c98a13b
Reduce dependencies (#16)
* remove magrittr

* remove fs

* update urls
