+++ R/octolog-package.R
@@ -3,7 +3,6 @@
 
 ## usethis namespace: start
 #' @importFrom glue glue
-#' @importFrom magrittr %>%
 #' @importFrom rlang %||%
 ## usethis namespace: end
 NULL
