+++ dev/pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 2.14.2
-pkgdown: 2.0.2
+pkgdown: 2.0.5
 pkgdown_sha: ~
 articles:
   octolog: octolog.html
-last_built: 2022-02-19T14:18Z
+last_built: 2022-06-25T17:49Z
 urls:
   reference: https://assignuser.github.io/octolog/reference
   article: https://assignuser.github.io/octolog/articles
