+++ CITATION.cff
@@ -1,5 +1,5 @@
 # -----------------------------------------------------------
-# CITATION file created with {cffr} R package, v0.2.1
+# CITATION file created with {cffr} R package, v0.2.2
 # See also: https://docs.ropensci.org/cffr/
 # -----------------------------------------------------------
  
@@ -7,12 +7,12 @@ cff-version: 1.2.0
 message: 'To cite package "octolog" in publications use:'
 type: software
 license: MIT
-title: 'octolog: Better ''Github Action'' Logging'
-version: 0.1.1
+title: 'octolog: Better ''GitHub Actions'' Logging'
+version: 0.1.1.9000
 doi: 10.5281/zenodo.5948954
-abstract: Provides an API for 'Github' <https://github.com> workflow commands and
+abstract: Provides an API for 'GitHub' <https://github.com> workflow commands and
   makes it possible to signal conditions from within R that will create annotations
-  when used within 'Github Actions' <https://github.com/features/actions> but raise
+  when used within 'GitHub Actions' <https://github.com/features/actions> but raise
   normal R conditions when used interactively.
 authors:
 - family-names: Wujciak-Jens
@@ -21,7 +21,7 @@ authors:
   orcid: https://orcid.org/0000-0002-7281-3989
 preferred-citation:
   type: manual
-  title: 'octolog: Better ''Github Action'' Logging'
+  title: 'octolog: Better ''GitHub Actions'' Logging'
   authors:
   - family-names: Wujciak-Jens
     given-names: Jacob
@@ -32,7 +32,7 @@ preferred-citation:
   url: https://doi.org/10.5281/zenodo.5948954
   doi: 10.5281/zenodo.5948954
 repository-code: https://github.com/assignUser/octolog
-url: https://assignuser.github.io/octolog/
+url: https://jacob.wujciak.de/octolog/
 contact:
 - family-names: Wujciak-Jens
   given-names: Jacob
@@ -40,6 +40,8 @@ contact:
   orcid: https://orcid.org/0000-0002-7281-3989
 keywords:
 - actions
+- ci
+- r
 - rstats
 - workflows
 references:
@@ -55,21 +57,6 @@ references:
   url: https://CRAN.R-project.org/package=cli
   version: '>= 3.0.0'
 - type: software
-  title: fs
-  abstract: 'fs: Cross-Platform File System Operations Based on ''libuv'''
-  notes: Imports
-  authors:
-  - family-names: Hester
-    given-names: Jim
-  - family-names: Wickham
-    given-names: Hadley
-    email: hadley@rstudio.com
-  - family-names: Csárdi
-    given-names: Gábor
-    email: csardi.gabor@gmail.com
-  year: '2022'
-  url: https://CRAN.R-project.org/package=fs
-- type: software
   title: glue
   abstract: 'glue: Interpreted String Literals'
   notes: Imports
@@ -84,31 +71,6 @@ references:
   year: '2022'
   url: https://CRAN.R-project.org/package=glue
 - type: software
-  title: magrittr
-  abstract: 'magrittr: A Forward-Pipe Operator for R'
-  notes: Imports
-  authors:
-  - family-names: Bache
-    given-names: Stefan Milton
-    email: stefan@stefanbache.dk
-  - family-names: Wickham
-    given-names: Hadley
-    email: hadley@rstudio.com
-  year: '2022'
-  url: https://CRAN.R-project.org/package=magrittr
-- type: software
-  title: openssl
-  abstract: 'openssl: Toolkit for Encryption, Signatures and Certificates Based on
-    OpenSSL'
-  notes: Imports
-  authors:
-  - family-names: Ooms
[…]
