+++ dev/pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.5
 pkgdown_sha: ~
 articles:
   octolog: octolog.html
-last_built: 2022-06-25T18:34Z
+last_built: 2022-06-25T18:37Z
 urls:
   reference: https://assignuser.github.io/octolog/reference
   article: https://assignuser.github.io/octolog/articles
