### 2022-06-25 #502 @assignUser
Thanks again to both of your for the informative reviews! It has been sometime but I managed to carve out some time to work on {octolog}. On that topic, @annakrystalli is there a deadline for me?

@pat-s I have removed {fs} and {magrittr}. {rlang} has to stay as it is a (hidden) dependency for [`cli_abort`](https://github.com/r-lib/cli/blob/main/R/rlang.R#L43) etc..
This also means that my actual dependencies are {cli} and {withr} with {glue} and {rlang} being "free" via {cli}. With these dependencies being very lightweight, without system dependencies and made for use in packages I don't think it is worth it to remove anymore dependencies (which would basically mean re-writing in base R 🙀 ).

While I am not entirely sure that {fs} is really such an issue (the compilation takes < 30s / use rspm & cache in CI!) it did not require big changes. {magrittr} on the other hand was completely superfluous and just added because I like how the code looks using it 😁.

@annakrystalli I removed most of the {lintr} issues but have too keep some of the line length ones as those are links and I have not found a way to add a line break [without breaking](https://github.com/assignUser/octolog/runs/7055289264?check_suite_focus=true#step:6:88) the link.

With the "technical" issues addressed I will work on improving the documentation next!
