### 2022-06-25 #502 @annakrystalli
Nice work @assignUser! While we suggest a response within a couple of weeks of reviews being submitted, there isn't a hard deadline really to when all comments must be resolved. The work required to resolve comments varies from package to package and we all understand having to carve out time can be hard especially during summer holiday time! So don't worry too much about a given deadline. Look forward to seeing the documentation! 😃
