### 2022-06-28 #502 @pat-s
> While I am not entirely sure that {fs} is really such an issue (the compilation takes < 30s / use rspm & cache in CI!) it did not require big changes. {magrittr} on the other hand was completely superfluous and just added because I like how the code looks using it 😁.

I also like to use both in scripting, though pkg dev is usually a different field. It's also not only about installation time/effort on CI but also on various systems (a system dep can be a blocker for some people in some restricted environments as they cannot install anything themselves - as a system engineer/admin I face this situation quite often). This reasoning does not really apply for your package as it is aimed at CI in the first place, but I'd would just call it "good practice" to think about these things and try to avoid such "convenience" packages in pkg dev if there are (robust) base R solutions - especially if the transition is cheap. Sometimes there are not or very complicated workarounds only, so it always differs...

(sorry for the wall of text, just some personal thoughts, it's ofc up to you what you do WRT to deps in the end :) )
