### 2023-02-03 #502 @maurolepore
Dear @assignUser, my EiC rotation just started and I'm checking the status of every open issue. Any updates?
