+++ .github/workflows/test-coverage.yaml
@@ -2,9 +2,8 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master]
+    branches: [main]
   pull_request:
-    branches: [main, master]
 
 name: test-coverage
 
@@ -15,7 +14,7 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
 
       - uses: r-lib/actions/setup-r@v2
         with:
@@ -26,5 +25,8 @@ jobs:
           extra-packages: any::covr
 
       - name: Test coverage
-        run: covr::codecov()
+        env:
+          COVERALLS_TOKEN: ${{ secrets.COVERALLS_REPO_TOKEN }}
+          CI_SERVICE: gha
+        run: covr::coveralls()
         shell: Rscript {0}
