### 2023-05-31 f17ff3a
Use coveralls instead of codecov (#21)
* Update test-coverage.yaml

* update test snapshot

* use coveralls action to submit

* add xml2 for conversion

* update badge
