+++ dev/pkgdown.yml
@@ -1,9 +1,9 @@
-pandoc: 2.14.2
-pkgdown: 2.0.5
+pandoc: 2.19.2
+pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   octolog: octolog.html
-last_built: 2022-06-25T18:37Z
+last_built: 2023-05-31T02:57Z
 urls:
   reference: https://assignuser.github.io/octolog/reference
   article: https://assignuser.github.io/octolog/articles
