### 2024-02-12 #502 @assignUser
I really hoped I would have more time to get this up to speed but at the moment that just hasn't worked out. I do appreciate the work everyone has put in so far, of course especially @pat-s and @annakrystalli as reviewers. You suggestions are very good and gave a good perspective and I hope to implement them at some point.

With my time constriants even if the package get's accepted it would be defacto abandond from the get go and I don't think that's the purpose of this process. For that reason I am going to clsoe this.

Thank you again everyone ❤️
